+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 14.3.0 Features
 
 - `[Calendar]` Add a setting for calendar to show and hide the legend. ([EP#6533](https://github.com/infor-design/enterprise/issues/6533))
+- `[Modal]` Update example page to test fullsize. ([#1358](https://github.com/infor-design/enterprise-ng/issues/1358))
 
 ### 14.3.0 Fixes